add parallel scrape database.
