+++ polo_trade_history/issue
@@ -1,0 +1,1 @@
+line 88 in utils will crash under python3, which the dict no longer have iteritems().
