+++ polo_trade_history/makefile
@@ -1,5 +1,5 @@
 all:
-	python get_trade_history.py
+	python get_trade_history.py pair.rc
 
 clean:
 	rm -rf log polo_history_data
